+++ src/typings.d.ts
@@ -1,5 +1,3 @@
-type Omit<T, K> = Pick<T, Exclude<keyof T, K>>
-
 declare namespace SwitchDo {
   interface Base {
     varName: string;